+++ index.ts
@@ -1,6 +1,7 @@
 import type { App, Plugin } from 'vue'
 import type { Router } from 'vue-router'
 import BeforeMiddleware from './src/middleware/before'
+import TitleManager from './src/middleware/title'
 import type { Guard } from './src/middleware/before'
 
 export { RouteGroup, Route } from './src/routing/Router'
@@ -10,6 +11,7 @@ export type { Guard, GuardContext, RouteRedirect } from './src/middleware/before
 export interface EasyRouteOptions {
   router: Router
   guards?: Guard[]
+  baseTitle?: string
 }
 
 const install: Plugin = (_app: App, options: EasyRouteOptions) => {
@@ -23,8 +25,10 @@ const install: Plugin = (_app: App, options: EasyRouteOptions) => {
 
   const router = options.router
   const guards = options.guards || []
+  const baseTitle = options.baseTitle || ''
 
   new BeforeMiddleware(router, guards)
+  new TitleManager(router, baseTitle)
 }
 
 export default install