+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Changelog
 
-### 2018-12-27 (2.0.8):
+### 2018-12-27 (1.0.2):
+
+- Feature: Router now sets page title on navigation
+
+### 2018-12-27 (1.0.1):
 
 - Bug fix: RouterGroup now checks for existance of the routes property
 - Refactor: Package description
 
-### 2018-11-30 (2.0.7):
+### 2018-11-30 (1.0.0):
 
 - Improved documentation
 - Strip name on parent routes